remove weird module wrapper
