--- index.mjs
@@ -1,38 +1,0 @@
-import createBareServer from "@tomphttp/bare-server-node";
-import http from "node:http";
-
-const httpServer = http.createServer();
-
-const bareServer = createBareServer("/", {
-  logErrors: false,
-  localAddress: undefined,
-  maintainer: {
-    email: "tomphttp@sys32.dev",
-    website: "https://github.com/tomphttp/",
-  },
-});
-
-httpServer.on("request", (req, res) => {
-  if (bareServer.shouldRoute(req)) {
-    bareServer.routeRequest(req, res);
-  } else {
-    res.writeHead(400);
-    res.end("Not found.");
-  }
-});
-
-httpServer.on("upgrade", (req, socket, head) => {
-  if (bareServer.shouldRoute(req)) {
-    bareServer.routeUpgrade(req, socket, head);
-  } else {
-    socket.end();
-  }
-});
-
-httpServer.on("listening", () => {
-  console.log("HTTP server listening");
-});
-
-httpServer.listen({
-  port: process.env.PORT || 6969,
-});
