+++ package.json
@@ -10,10 +10,11 @@
   "name": "uv-app",
   "version": "1.0.0",
   "description": "Ultraviolet instance",
+  "type": "module",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node ."
+    "start": "node index.js"
   },
   "keywords": [
     "proxy"
